fix[api]: propagate diagnostics through model loading (QVAC-21926)
