+++ packages/sdk/test/unit/resource-collector.test.ts
@@ -5,11 +5,7 @@ import {
   initializeWorkerResourceCollector
 } from '@/server/bare/resources/worker-collector'
 import { handleGetSystemResources } from '@/server/rpc/handlers/get-system-resources'
-import {
-  backendProbeResultSchema,
-  getSystemResourcesResponseSchema,
-  inferenceBackendDiagnosticsSchema
-} from '@/schemas/system-resources'
+import { getSystemResourcesResponseSchema } from '@/schemas/system-resources'
 
 const drivers = {
   vulkan: false,
@@ -386,40 +382,4 @@ test('resource RPC reports failed metrics when the collector is unavailable', (t
   t.is(response.capabilities.gpus.status, 'failed')
   t.is(response.sample?.cpu.status, 'failed')
   t.execution(() => getSystemResourcesResponseSchema.parse(response))
-})
-
-test('backend diagnostics validate probe, selection, and fallback details', (t) => {
-  t.ok(
-    inferenceBackendDiagnosticsSchema.safeParse({
-      selectedBackend: 'llama.cpp-metal',
-      selectedDevice: 'gpu',
-      graphicsApi: 'metal',
-      driver: { name: 'Metal', version: '3' },
-      gpuId: 'gpu-opaque-1',
-      fallback: {
-        requestedBackend: 'llama.cpp-vulkan',
-        requestedDevice: 'gpu',
-        reason: 'Vulkan backend is unavailable'
-      },
-      probe: {
-        status: 'compatible',
-        backend: 'llama.cpp-metal'
-      }
-    }).success
-  )
-  t.absent(
-    backendProbeResultSchema.safeParse({
-      status: 'incompatible',
-      backend: ''
-    }).success,
-    'probe backend cannot be empty'
-  )
-  t.absent(
-    inferenceBackendDiagnosticsSchema.safeParse({
-      selectedBackend: 'llama.cpp-cpu',
-      selectedDevice: 'cpu',
-      fallback: {}
-    }).success,
-    'fallback requires a reason'
-  )
 })
